using new recipe model now instead of old one
