+++ app/Data/AppDatabaseContext.cs
@@ -60,7 +60,7 @@ public class AppDatabaseContext : DbContext
 
     public DbSet<UserModel> User { get; set; } = null!;
 
-    public DbSet<NewRecipeModel> Recipe { get; set; } = null!;
+    public DbSet<OldRecipeModel> Recipe { get; set; } = null!;
 
     public DbSet<IngredientModel> Ingredient { get; set; } = null!;
