+++ app/Models/RecipeIngredientModel.cs
@@ -13,7 +13,7 @@ public class RecipeIngredientModel
 
     public string IngredientCookingMethod { get; set; } = "";
 
-    public NewRecipeModel Recipe { get; set; } = null!;
+    public RecipeModel Recipe { get; set; } = null!;
 
     public IngredientModel Ingredient { get; set; } = null!;
 }
